+++ sdk/logic/arm-logic/src/operations/integrationAccountAgreements.ts
@@ -11,7 +11,7 @@ import { IntegrationAccountAgreements } from "../operationsInterfaces";
 import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers";
 import * as Parameters from "../models/parameters";
-import { LogicManagementClientContext } from "../logicManagementClientContext";
+import { LogicManagementClient } from "../logicManagementClient";
 import {
   IntegrationAccountAgreement,
   IntegrationAccountAgreementsListNextOptionalParams,
@@ -32,13 +32,13 @@ import {
 /** Class containing IntegrationAccountAgreements operations. */
 export class IntegrationAccountAgreementsImpl
   implements IntegrationAccountAgreements {
-  private readonly client: LogicManagementClientContext;
+  private readonly client: LogicManagementClient;
 
   /**
    * Initialize a new instance of the class IntegrationAccountAgreements class.
    * @param client Reference to the service client
    */
-  constructor(client: LogicManagementClientContext) {
+  constructor(client: LogicManagementClient) {
     this.client = client;
   }