+++ sdk/storage/storage-queue/test/aborter.spec.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import * as assert from "assert";
+import { assert } from "chai";
 import { AbortController } from "@azure/abort-controller";
 
 import { QueueClient } from "../src/QueueClient";